added saving language preference to database, added basic default user functionality, removed Log In As Guest

+++ src/pl/edu/agh/fragments/HelpElementDescriptionFragment.java
@@ -15,7 +15,7 @@ public class HelpElementDescriptionFragment  extends AbstractDescriptionFragment
 
     private HelpElement displayedHelpElement;
 
-    public static HelpElementDescriptionFragment newInstance(HelpElement helpElement, int index) {
+    public static HelpElementDescriptionFragment newInstance(HelpElement helpElement, long index) {
         HelpElementDescriptionFragment fragment = new HelpElementDescriptionFragment();
         fragment.setInitialArguments(index, helpElement);
         return fragment;
